+++ applications/services/gui/gui_i.h
@@ -59,7 +59,7 @@ struct Gui {
     // Input
     FuriMessageQueue* input_queue;
     FuriPubSub* input_events;
-    uint8_t ongoing_input;
+    uint32_t ongoing_input;
     ViewPort* ongoing_input_view_port;
 };
